Crate image_moments

Efficient and compile-time checked calculations of image moments

Structs

Central

The central moments of an image or contour which are translational invariant.

Index

An index specified at compile time.

NormalizedCentral

The normalized central moments of an image or contour which are translational and scale invariant.

Order

An order of image moments specified at compile time.

Spatial

The raw, spatial moments of an image or contour.

Traits

Moments

A generalization over different moments.

Point

A generalization over different possible representations of points. Implementing this trait on custom structs ensure their seamless usage with this crate.

Scalar

The type of floating-point number used to calculate the image moments.

SupportedIndex

An marker trait indicating a compile-time index is valid for a specific order. This trait is sealed.

SupportedOrder

An order currently supported by this crate for calculating image moments. The trait is sealed and the actual implementation details hidden from the user.