Trait image_moments::Point

pub trait Point<S: Scalar>: Clone {
    fn x(&self) -> S;

    fn y(&self) -> S;
}

A generalization over different possible representations of points. Implementing this trait on custom structs ensure their seamless usage with this crate.

Required methods

fn x(&self) -> S

The x position of the point casted into the required resolution.

fn y(&self) -> S

The y position of the point casted into the required resolution.