Trait image_moments::Scalar

pub trait Scalar: Debug + Copy + PartialEq + PartialOrd + AddAssign + MulAssign + Neg<Output = Self> + Add<Self, Output = Self> + Sub<Self, Output = Self> + Mul<Self, Output = Self> + Div<Self, Output = Self> {
    const EPSILON: Self;
    const ZERO: Self;
    const ONE: Self;
    const TWO: Self;
    const THREE: Self;
    const F1_2: Self;
    const F1_6: Self;
    const F1_12: Self;
    const F1_20: Self;
    const F1_24: Self;
    const F1_60: Self;

    fn abs(self) -> Self;

    fn mul_add(self, a: Self, b: Self) -> Self;

    fn powi(self, n: i32) -> Self;

    fn copysign(self, sign: Self) -> Self;

    fn sqrt(self) -> Self;
}

The type of floating-point number used to calculate the image moments.

Associated Constants

const EPSILON: Self

The smallest incremental step of this floating-point number.

const ZERO: Self

The value of the scalar matching 0.0.

const ONE: Self

The value of the scalar matching 1.0.

const TWO: Self

The value of the scalar matching 2.0.

const THREE: Self

The value of the scalar matching 3.0.

const F1_2: Self

The value of the scalar matching 1/2.

const F1_6: Self

The value of the scalar matching 1/6.

const F1_12: Self

The value of the scalar matching 1/12.

const F1_20: Self

The value of the scalar matching 1/20.

const F1_24: Self

The value of the scalar matching 1/24.

const F1_60: Self

The value of the scalar matching 1/60.

Required methods

fn abs(self) -> Self

Computes the absolute value of self

fn mul_add(self, a: Self, b: Self) -> Self

Fused multiply-add. Computes (self * a) + b with only one rounding error, yielding a more accurate result than an unfused multiply-add.

fn powi(self, n: i32) -> Self

Raises a number to an integer power.

fn copysign(self, sign: Self) -> Self

Returns a number composed of the magnitude of self and the sign of sign.

fn sqrt(self) -> Self

Calculate the square root.